+++ client/questmark/src/App.css
@@ -20,35 +20,3 @@
 .App-link {
   color: #09d3ac;
 }
-
-.mainContainer{
-  display: flex;
-  flex-direction: row;
-  min-height: 80vh;
-  flex-wrap: wrap;
-  justify-content: center
-}
-
-.noForms{
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  height: 10vh;  
-  background-color: #254D85;
-  margin: 10vh;
-  border-color: #769ACC;
-  border-width: 1px;
-  border-style: solid;
-  color : white;
-  text-align: center;
-  font-size: 20px;
-  justify-content: center;
-  align-items: center
-}
-
-.createForm{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center
-}
