+++ client/questmark/src/components/Header.js
@@ -17,12 +17,16 @@ export default class Header extends Component {
             <header className="mainHeader">
                 Questmark
                 <div className="options">
+                    {this.props.ordernar === true ?
                     <select name="ordernar">
                     <option value="" selected>Ordenar por :</option>
                         <option value="alfabetica">Ordem Alfabética [A-Z]</option>
                         <option value="criacao">Data de criação</option>
                         <option value="edicao">Data de modificação</option>
                     </select>
+                    :
+                    <div/>
+                    }
                     <p>Ajuda</p>
                     <p>Opções</p>
                     <p>Log Out</p>
