+++ client/questmark/src/pages/CreateForm.css
@@ -1,35 +1,33 @@
 .formContainer{
     display: flex;
     flex-direction: column;
-    background-color: black;
     min-height: 90vh;
 }
 
 .formsField{
     display: flex;    
     min-height: 70vh;
-    background-color: pink;
 }
 .inputMD{
     display: flex;
     flex-direction: column;
     flex: 1;
-    background-color: blueviolet;
     padding : 3vh;
     border-right-width: 1px;
     border-right-color: black;
     border-right-style: solid;
+    margin-top : 1vh;
 }
 
 .outputHTML{
     display: flex;
     flex-direction: column;
     flex : 1;
-    background-color: green;
     padding : 3vh;
     border-left-width: 1px;
     border-left-color: black;
     border-left-style: solid;
+    margin-top : 1vh;
 }
 
 .formsField textarea{
@@ -39,17 +37,16 @@
 .formName{  
     display: flex;
     flex-direction: row;
-    background-color: yellow;
     min-height: 7vh;
     align-items: center;
     justify-content: center;
+    margin-top : 1vh;
 }
 
 .formName input{
     min-height: 5vh;
     min-width: 20%;
     margin-left: 2vh;
-    border-radius: 10px;
 }
 
 .buttons{
